virttest: Fix giant 'from __init__ import ping' issue
Fix a circular dependency from the utils_test module
once for all by moving raw_ping and ping to virttest.utils_net
and just letting references in the original places so
tests that use utils_test.raw_ping and utils_test.ping
would still work.
