+++ etc/avocado/conf.d/virt-test.conf
@@ -31,7 +31,7 @@ qemu_dst_bin =
 # Accelerator used to run qemu (kvm or tcg)
 accel = kvm
 # Nettype (bridge, user, none)
-nettype = user
+nettype = bridge
 # Bridge name to be used if you select bridge as a nettype
 netdst = virbr0
 # Whether to enable vhost for qemu (on/off/force). Depends on nettype=bridge
@@ -59,9 +59,9 @@ malloc_perturb = yes
 # Test connect URI for libvirt (qemu:///system', 'lxc:///')
 connect_uri = qemu:///session
 # Install the guest using import method before the tests are run
-install_guest = False
+install_guest = True
 # Remove the guest from libvirt. This will not delete the guest's disk file
-remove_guest = False
+remove_guest = True
 [virt_test.debug]
 # Don't clean up tmp files or VM processes at the end of a virt-test execution
 no_cleanup = False
